+++ tgui/packages/tgui/interfaces/VorePanel/VoreUserPreferences.tsx.tsx
@@ -1,0 +1,438 @@
+import { BooleanLike } from 'common/react';
+
+import { useBackend } from '../../backend';
+import { Box, Button, Divider, Flex, Section } from '../../components';
+import { prefData } from './types';
+import { VoreUserPreferencesAesthetic } from './VoreUserPreferencesTabs/VoreUserPreferencesAesthetic ';
+import { VoreUserPreferencesMechanical } from './VoreUserPreferencesTabs/VoreUserPreferencesMechanical ';
+
+export const VoreUserPreferences = (props: {
+  prefs: prefData;
+  show_pictures: BooleanLike;
+}) => {
+  const { act } = useBackend();
+
+  const { prefs, show_pictures } = props;
+
+  const {
+    digestable,
+    devourable,
+    resizable,
+    feeding,
+    absorbable,
+    digest_leave_remains,
+    allowmobvore,
+    permit_healbelly,
+    show_vore_fx,
+    can_be_drop_prey,
+    can_be_drop_pred,
+    allow_inbelly_spawning,
+    allow_spontaneous_tf,
+    step_mechanics_active,
+    pickup_mechanics_active,
+    noisy,
+    drop_vore,
+    stumble_vore,
+    slip_vore,
+    throw_vore,
+    food_vore,
+    digest_pain,
+    nutrition_message_visible,
+    weight_message_visible,
+    eating_privacy_global,
+  } = prefs;
+
+  const preferences = {
+    digestion: {
+      action: 'toggle_digest',
+      test: digestable,
+      tooltip: {
+        main: "This button is for those who don't like being digested. It can make you undigestable.",
+        enable: 'Click here to allow digestion.',
+        disable: 'Click here to prevent digestion.',
+      },
+      content: {
+        enabled: 'Digestion Allowed',
+        disabled: 'No Digestion',
+      },
+    },
+    absorbable: {
+      action: 'toggle_absorbable',
+      test: absorbable,
+      tooltip: {
+        main: "This button allows preds to know whether you prefer or don't prefer to be absorbed.",
+        enable: 'Click here to allow being absorbed.',
+        disable: 'Click here to disallow being absorbed.',
+      },
+      content: {
+        enabled: 'Absorption Allowed',
+        disabled: 'No Absorption',
+      },
+    },
+    devour: {
+      action: 'toggle_devour',
+      test: devourable,
+      tooltip: {
+        main: 'This button is to toggle your ability to be devoured by others.',
+        enable: 'Click here to allow being devoured.',
+        disable: 'Click here to prevent being devoured.',
+      },
+      content: {
+        enabled: 'Devouring Allowed',
+        disabled: 'No Devouring',
+      },
+    },
+    mobvore: {
+      action: 'toggle_mobvore',
+      test: allowmobvore,
+      tooltip: {
+        main: "This button is for those who don't like being eaten by mobs.",
+        enable: 'Click here to allow being eaten by mobs.',
+        disable: 'Click here to prevent being eaten by mobs.',
+      },
+      content: {
+        enabled: 'Mobs eating you allowed',
+        disabled: 'No Mobs eating you',
+      },
+    },
+    feed: {
+      action: 'toggle_feed',
+      test: feeding,
+      tooltip: {
+        main: 'This button is to toggle your ability to be fed to or by others vorishly.',
+        enable: 'Click here to allow being fed to/by other people.',
+        disable: 'Click here to prevent being fed to/by other people.',
+      },
+      content: {
+        enabled: 'Feeding Allowed',
+        disabled: 'No Feeding',
+      },
+    },
+    healbelly: {
+      action: 'toggle_healbelly',
+      test: permit_healbelly,
+      tooltip: {
+        main:
+          "This button is for those who don't like healbelly used on them as a mechanic." +
+          ' It does not affect anything, but is displayed under mechanical prefs for ease of quick checks.',
+        enable: 'Click here to allow being heal-bellied.',
+        disable: 'Click here to prevent being heal-bellied.',
+      },
+      content: {
+        enabled: 'Heal-bellies Allowed',
+        disabled: 'No Heal-bellies',
+      },
+    },
+    dropnom_prey: {
+      action: 'toggle_dropnom_prey',
+      test: can_be_drop_prey,
+      tooltip: {
+        main:
+          'This toggle is for spontaneous, environment related vore' +
+          ' as prey, including drop-noms, teleporters, etc.',
+        enable: 'Click here to allow being spontaneous prey.',
+        disable: 'Click here to prevent being spontaneous prey.',
+      },
+      content: {
+        enabled: 'Spontaneous Prey Enabled',
+        disabled: 'Spontaneous Prey Disabled',
+      },
+    },
+    dropnom_pred: {
+      action: 'toggle_dropnom_pred',
+      test: can_be_drop_pred,
+      tooltip: {
+        main:
+          'This toggle is for spontaneous, environment related vore' +
+          ' as a predator, including drop-noms, teleporters, etc.',
+        enable: 'Click here to allow being spontaneous pred.',
+        disable: 'Click here to prevent being spontaneous pred.',
+      },
+      content: {
+        enabled: 'Spontaneous Pred Enabled',
+        disabled: 'Spontaneous Pred Disabled',
+      },
+    },
+    toggle_drop_vore: {
+      action: 'toggle_drop_vore',
+      test: drop_vore,
+      tooltip: {
+        main:
+          'Allows for dropnom spontaneous vore to occur. ' +
+          'Note, you still need spontaneous vore pred and/or prey enabled.',
+        enable: 'Click here to allow for dropnoms.',
+        disable: 'Click here to disable dropnoms.',
+      },
+      content: {
+        enabled: 'Drop Noms Enabled',
+        disabled: 'Drop Noms Disabled',
+      },
+    },
+    toggle_slip_vore: {
+      action: 'toggle_slip_vore',
+      test: slip_vore,
+      tooltip: {
+        main:
+          'Allows for slip related spontaneous vore to occur. ' +
+          'Note, you still need spontaneous vore pred and/or prey enabled.',
+        enable: 'Click here to allow for slip vore.',
+        disable: 'Click here to disable slip vore.',
+      },
+      content: {
+        enabled: 'Slip Vore Enabled',
+        disabled: 'Slip Vore Disabled',
+      },
+    },
+    toggle_stumble_vore: {
+      action: 'toggle_stumble_vore',
+      test: stumble_vore,
+      tooltip: {
+        main:
+          'Allows for stumble related spontaneous vore to occur. ' +
+          ' Note, you still need spontaneous vore pred and/or prey enabled.',
+        enable: 'Click here to allow for stumble vore.',
+        disable: 'Click here to disable stumble vore.',
+      },
+      content: {
+        enabled: 'Stumble Vore Enabled',
+        disabled: 'Stumble Vore Disabled',
+      },
+    },
+    toggle_throw_vore: {
+      action: 'toggle_throw_vore',
+      test: throw_vore,
+      tooltip: {
+        main:
+          'Allows for throw related spontaneous vore to occur. ' +
+          ' Note, you still need spontaneous vore pred and/or prey enabled.',
+        enable: 'Click here to allow for throw vore.',
+        disable: 'Click here to disable throw vore.',
+      },
+      content: {
+        enabled: 'Throw Vore Enabled',
+        disabled: 'Throw Vore Disabled',
+      },
+    },
+    toggle_food_vore: {
+      action: 'toggle_food_vore',
+      test: food_vore,
+      tooltip: {
+        main:
+          'Allows for food related spontaneous vore to occur. ' +
+          ' Note, you still need spontaneous vore pred and/or prey enabled.',
+        enable: 'Click here to allow for food vore.',
+        disable: 'Click here to disable food vore.',
+      },
+      content: {
+        enabled: 'Food Vore Enabled',
+        disabled: 'Food Vore Disabled',
+      },
+    },
+    toggle_digest_pain: {
+      action: 'toggle_digest_pain',
+      test: digest_pain,
+      tooltip: {
+        main:
+          'Allows for pain messages to show when being digested. ' +
+          ' Can be toggled off to disable pain messages.',
+        enable: 'Click here to allow for digestion pain.',
+        disable: 'Click here to disable digestion pain.',
+      },
+      content: {
+        enabled: 'Digestion Pain Enabled',
+        disabled: 'Digestion Pain Disabled',
+      },
+    },
+    inbelly_spawning: {
+      action: 'toggle_allow_inbelly_spawning',
+      test: allow_inbelly_spawning,
+      tooltip: {
+        main:
+          'This toggle is ghosts being able to spawn in one of your bellies.' +
+          ' You will have to confirm again when they attempt to.',
+        enable: 'Click here to allow prey to spawn in you.',
+        disable: 'Click here to prevent prey from spawning in you.',
+      },
+      content: {
+        enabled: 'Inbelly Spawning Allowed',
+        disabled: 'Inbelly Spawning Forbidden',
+      },
+    },
+    noisy: {
+      action: 'toggle_noisy',
+      test: noisy,
+      tooltip: {
+        main: 'Toggle audible hunger noises.',
+        enable: 'Click here to turn on hunger noises.',
+        disable: 'Click here to turn off hunger noises.',
+      },
+      content: {
+        enabled: 'Hunger Noises Enabled',
+        disabled: 'Hunger Noises Disabled',
+      },
+    },
+    resize: {
+      action: 'toggle_resize',
+      test: resizable,
+      tooltip: {
+        main: 'This button is to toggle your ability to be resized by others.',
+        enable: 'Click here to allow being resized.',
+        disable: 'Click here to prevent being resized.',
+      },
+      content: {
+        enabled: 'Resizing Allowed',
+        disabled: 'No Resizing',
+      },
+    },
+    steppref: {
+      action: 'toggle_steppref',
+      test: step_mechanics_active,
+      tooltip: {
+        main: '',
+        enable:
+          'You will not participate in step mechanics.' +
+          ' Click to enable step mechanics.',
+        disable:
+          'This setting controls whether or not you participate in size-based step mechanics.' +
+          ' Includes both stepping on others, as well as getting stepped on. Click to disable step mechanics.',
+      },
+      content: {
+        enabled: 'Step Mechanics Enabled',
+        disabled: 'Step Mechanics Disabled',
+      },
+    },
+    vore_fx: {
+      action: 'toggle_fx',
+      test: show_vore_fx,
+      tooltip: {
+        main: '',
+        enable:
+          'Regardless of Predator Setting, you will not see their FX settings.' +
+          ' Click this to enable showing FX.',
+        disable:
+          'This setting controls whether or not a pred is allowed to mess with your HUD and fullscreen overlays.' +
+          ' Click to disable all FX.',
+      },
+      content: {
+        enabled: 'Show Vore FX',
+        disabled: 'Do Not Show Vore FX',
+      },
+    },
+    remains: {
+      action: 'toggle_leaveremains',
+      test: digest_leave_remains,
+      tooltip: {
+        main: '',
+        enable:
+          'Regardless of Predator Setting, you will not leave remains behind.' +
+          ' Click this to allow leaving remains.',
+        disable:
+          'Your Predator must have this setting enabled in their belly modes to allow remains to show up,' +
+          ' if they do not, they will not leave your remains behind, even with this on. Click to disable remains.',
+      },
+      content: {
+        enabled: 'Allow Leaving Remains',
+        disabled: 'Do Not Allow Leaving Remains',
+      },
+    },
+    pickuppref: {
+      action: 'toggle_pickuppref',
+      test: pickup_mechanics_active,
+      tooltip: {
+        main: '',
+        enable:
+          'You will not participate in pick-up mechanics.' +
+          ' Click this to allow picking up/being picked up.',
+        disable:
+          'Allows macros to pick you up into their hands, and you to pick up micros.' +
+          ' Click to disable pick-up mechanics.',
+      },
+      content: {
+        enabled: 'Pick-up Mechanics Enabled',
+        disabled: 'Pick-up Mechanics Disabled',
+      },
+    },
+    spontaneous_tf: {
+      action: 'toggle_allow_spontaneous_tf',
+      test: allow_spontaneous_tf,
+      tooltip: {
+        main:
+          'This toggle is for spontaneous or environment related transformation' +
+          ' as a victim, such as via chemicals.',
+        enable: 'Click here to allow being spontaneously transformed.',
+        disable: 'Click here to disable being spontaneously transformed.',
+      },
+      content: {
+        enabled: 'Spontaneous TF Enabled',
+        disabled: 'Spontaneous TF Disabled',
+      },
+    },
+    examine_nutrition: {
+      action: 'toggle_nutrition_ex',
+      test: nutrition_message_visible,
+      tooltip: {
+        main: '',
+        enable: 'Click here to enable nutrition messages.',
+        disable: 'Click here to disable nutrition messages.',
+      },
+      content: {
+        enabled: 'Examine Nutrition Messages Active',
+        disabled: 'Examine Nutrition Messages Inactive',
+      },
+    },
+    examine_weight: {
+      action: 'toggle_weight_ex',
+      test: weight_message_visible,
+      tooltip: {
+        main: '',
+        enable: 'Click here to enable weight messages.',
+        disable: 'Click here to disable weight messages.',
+      },
+      content: {
+        enabled: 'Examine Weight Messages Active',
+        disabled: 'Examine Weight Messages Inactive',
+      },
+    },
+    eating_privacy_global: {
+      action: 'toggle_global_privacy',
+      test: eating_privacy_global,
+      tooltip: {
+        main:
+          'Sets default belly behaviour for vorebellies for announcing' +
+          ' ingesting or expelling prey' +
+          ' Overwritten by belly-specific preferences if set.',
+        enable: ' Click here to turn your messages subtle',
+        disable: ' Click here to turn your  messages loud',
+      },
+      content: {
+        enabled: 'Global Vore Privacy: Subtle',
+        disabled: 'Global Vore Privacy: Loud',
+      },
+    },
+  };
+
+  return (
+    <Box nowrap>
+      <VoreUserPreferencesMechanical
+        show_pictures={show_pictures}
+        preferences={preferences}
+      />
+      <VoreUserPreferencesAesthetic preferences={preferences} />
+      <Divider />
+      <Section>
+        <Flex spacing={1}>
+          <Flex.Item basis="49%">
+            <Button fluid icon="save" onClick={() => act('saveprefs')}>
+              Save Prefs
+            </Button>
+          </Flex.Item>
+          <Flex.Item basis="49%" grow={1}>
+            <Button fluid icon="undo" onClick={() => act('reloadprefs')}>
+              Reload Prefs
+            </Button>
+          </Flex.Item>
+        </Flex>
+      </Section>
+    </Box>
+  );
+};
