+++ tgui/packages/tgui/interfaces/VorePanelExport/VorePanelExportBellyString.tsx
@@ -1,183 +1,9 @@
-import { BooleanLike } from 'common/react';
-
-import { useBackend } from '../backend';
-import { Button, Section } from '../components';
-import { Window } from '../layouts';
-
-const ModeSpan = {
-  Hold: '<span class="badge text-bg-secondary">Hold</span>',
-  Digest: '<span class="badge text-bg-danger">Digest</span>',
-  Absorb: '<span class="badge text-bg-light">Absorb</span>',
-  Drain: '<span class="badge text-bg-warning">Drain</span>',
-  Selective: '<span class="badge text-bg-warning">Selective</span>',
-  Unabsorb: '<span class="badge text-bg-light">Unabsorb</span>',
-  Heal: '<span class="badge text-bg-success">Heal</span>',
-  Shrink: '<span class="badge text-bg-info">Shrink</span>',
-  Grow: '<span class="badge text-bg-info">Grow</span>',
-  'Size Steal': '<span class="badge text-bg-info">Size Steal</span>',
-  'Encase In Egg': '<span class="badge text-bg-primary">Encase In Egg</span>',
-};
-
-const ItemModeSpan = {
-  Hold: '<span class="badge text-bg-secondary">Item: Hold</span>',
-  'Digest (Food Only)':
-    '<span class="badge text-bg-danger">Item: Digest (Food Only)</span>',
-  Digest: '<span class="badge text-bg-danger">Item: Digest</span>',
-};
-
-const AddonIcon = {
-  Numbing: '',
-  Stripping: '',
-  'Leave Remains': '',
-  Muffles: 'bi-volume-mute',
-  'Affect Worn Items': '',
-  'Jams Sensors': 'bi-wifi-off',
-  'Complete Absorb': '',
-};
-
-const GetAddons = (addons: string[]) => {
-  let result: string[] = [];
-
-  addons?.forEach((addon) => {
-    result.push(
-      '<span class="badge text-bg-secondary"><i class="' +
-        AddonIcon[addon] +
-        '"></i>' +
-        addon +
-        '</span>',
-    );
-  });
-
-  if (result.length === 0) {
-    result.push('No Addons Set');
-  }
-
-  return result;
-};
-
-type Data = {
-  db_version: string;
-  db_repo: string;
-  mob_name: string;
-  bellies: Belly[];
-};
-
-type Belly = {
-  // General Information
-  name: string;
-  desc: string;
-  message_mode: BooleanLike;
-  absorbed_desc: string;
-  vore_verb: string;
-  release_verb: string;
-
-  // Controls
-  mode: string;
-  addons: string[];
-  item_mode: string;
-
-  // Options
-  digest_brute: number;
-  digest_burn: number;
-  digest_oxy: number;
-  digest_tox: number;
-  digest_clone: number;
-
-  can_taste: BooleanLike;
-  contaminates: BooleanLike;
-  contamination_flavor: string;
-  contamination_color: string;
-  nutrition_percent: number;
-  bulge_size: number;
-  display_absorbed_examine: BooleanLike;
-  save_digest_mode: BooleanLike;
-  emote_active: BooleanLike;
-  emote_time: number;
-  shrink_grow_size: number;
-  egg_type: string;
-  selective_preference: string;
-
-  // Messages
-  struggle_messages_outside: string[];
-  struggle_messages_inside: string[];
-  absorbed_struggle_messages_outside: string[];
-  absorbed_struggle_messages_inside: string[];
-  escape_attempt_messages_owner: string[];
-  escape_attempt_messages_prey: string[];
-  escape_messages_owner: string[];
-  escape_messages_prey: string[];
-  escape_messages_outside: string[];
-  escape_item_messages_owner: string[];
-  escape_item_messages_prey: string[];
-  escape_item_messages_outside: string[];
-  escape_fail_messages_owner: string[];
-  escape_fail_messages_prey: string[];
-  escape_attempt_absorbed_messages_owner: string[];
-  escape_attempt_absorbed_messages_prey: string[];
-  escape_absorbed_messages_owner: string[];
-  escape_absorbed_messages_prey: string[];
-  escape_absorbed_messages_outside: string[];
-  escape_fail_absorbed_messages_owner: string[];
-  escape_fail_absorbed_messages_prey: string[];
-  primary_transfer_messages_owner: string[];
-  primary_transfer_messages_prey: string[];
-  secondary_transfer_messages_owner: string[];
-  secondary_transfer_messages_prey: string[];
-  digest_chance_messages_owner: string[];
-  digest_chance_messages_prey: string[];
-  absorb_chance_messages_owner: string[];
-  absorb_chance_messages_prey: string[];
-  digest_messages_owner: string[];
-  digest_messages_prey: string[];
-  absorb_messages_owner: string[];
-  absorb_messages_prey: string[];
-  unabsorb_messages_owner: string[];
-  unabsorb_messages_prey: string[];
-  examine_messages: string[];
-  examine_messages_absorbed: string[];
-
-  // emote_list: any[];
-  emotes_digest;
-  emotes_hold;
-  emotes_holdabsorbed;
-  emotes_absorb;
-  emotes_heal;
-  emotes_drain;
-  emotes_steal;
-  emotes_egg;
-  emotes_shrink;
-  emotes_grow;
-  emotes_unabsorb;
-
-  // Sounds
-  is_wet: BooleanLike;
-  wet_loop: BooleanLike;
-  fancy_vore: BooleanLike;
-  vore_sound: string;
-  release_sound: string;
-
-  // Visuals (Vore FX)
-  disable_hud: BooleanLike;
-
-  // Interactions
-  escapable: BooleanLike;
-
-  escapechance: number;
-  escapechance_absorbed: number;
-  escapetime: number;
-
-  transferchance: number;
-  transferlocation: string;
-
-  transferchance_secondary: number;
-  transferlocation_secondary: string;
-
-  absorbchance: number;
-  digestchance: number;
-};
+import { ItemModeSpan, ModeSpan } from './constants';
+import { Belly } from './types';
+import { GetAddons } from './VorePanelExportBellyStringHelpers';
 
 // prettier-ignore
-const generateBellyString = (belly: Belly, index: number) => {
+export const generateBellyString = (belly: Belly, index: number) => {
   const {
     // General Information
     name,
@@ -768,129 +594,4 @@ const generateBellyString = (belly: Belly, index: number) => {
   result += '</div></div></div>';
 
   return result;
-};
-
-const getCurrentTimestamp = (): string => {
-  let now = new Date();
-  let hours = String(now.getHours());
-  if (hours.length < 2) {
-    hours = '0' + hours;
-  }
-  let minutes = String(now.getMinutes());
-  if (minutes.length < 2) {
-    minutes = '0' + minutes;
-  }
-  let dayofmonth = String(now.getDate());
-  if (dayofmonth.length < 2) {
-    dayofmonth = '0' + dayofmonth;
-  }
-  let month = String(now.getMonth() + 1); // 0-11
-  if (month.length < 2) {
-    month = '0' + month;
-  }
-  let year = String(now.getFullYear());
-
-  return (
-    ' ' +
-    year +
-    '-' +
-    month +
-    '-' +
-    dayofmonth +
-    ' (' +
-    hours +
-    ' ' +
-    minutes +
-    ')'
-  );
-};
-
-const downloadPrefs = (extension: string) => {
-  const { act, data } = useBackend<Data>();
-
-  const { db_version, db_repo, mob_name, bellies } = data;
-
-  let datesegment = getCurrentTimestamp();
-
-  let filename = mob_name + datesegment + extension;
-  let blob;
-
-  if (extension === '.html') {
-    let style = '<style>' + '</style>';
-
-    blob = new Blob(
-      [
-        '<!DOCTYPE html><html lang="en"><head>' +
-          '<meta charset="utf-8">' +
-          '<meta name="viewport" content="width=device-width, initial-scale=1">' +
-          '<title>' +
-          bellies.length +
-          ' Exported Bellies (DB_VER: ' +
-          db_repo +
-          '-' +
-          db_version +
-          ')</title>' +
-          '<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous">' +
-          '<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.9.1/font/bootstrap-icons.css">' +
-          style +
-          '</head><body class="py-4"><main><div class="container"><h2>Bellies of ' +
-          mob_name +
-          '</h2><p class="lead">Generated on: ' +
-          datesegment +
-          '</p><div class="accordion" id="accordionBellies">',
-      ],
-      {
-        type: 'text/html;charset=utf8',
-      },
-    );
-    bellies.forEach((belly, i) => {
-      blob = new Blob([blob, generateBellyString(belly, i)], {
-        type: 'text/html;charset=utf8',
-      });
-    });
-    blob = new Blob(
-      [
-        blob,
-        '</div>',
-        '<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-A3rJD856KowSb7dwlZdYEkO39Gagi7vIsF0jrRAoQmDKKtQBHUuLZ9AsSv4jD4Xa" crossorigin="anonymous"></script>',
-        '</div></main></body></html>',
-      ],
-      { type: 'text/html;charset=utf8' },
-    );
-  }
-
-  if (extension === '.vrdb') {
-    blob = new Blob([JSON.stringify(bellies)], { type: 'application/json' });
-  }
-
-  (window.navigator as any).msSaveOrOpenBlob(blob, filename);
-};
-
-export const VorePanelExport = () => {
-  return (
-    <Window width={790} height={560} theme="abstract">
-      <Window.Content>
-        <VorePanelExportContent />
-      </Window.Content>
-    </Window>
-  );
-};
-
-const VorePanelExportContent = (props) => {
-  const { act, data } = useBackend<Data>();
-
-  const { bellies } = data;
-
-  return (
-    <Section title="Vore Export Panel">
-      <Section title="Export">
-        <Button fluid icon="file-alt" onClick={() => downloadPrefs('.html')}>
-          Export (HTML)
-        </Button>
-        <Button fluid icon="file-alt" onClick={() => downloadPrefs('.vrdb')}>
-          Export (VRDB)
-        </Button>
-      </Section>
-    </Section>
-  );
 };
