+++ tgui/packages/tgui/interfaces/VorePanelExport/constants.ts
@@ -1,0 +1,30 @@
+export const ModeSpan = {
+  Hold: '<span class="badge text-bg-secondary">Hold</span>',
+  Digest: '<span class="badge text-bg-danger">Digest</span>',
+  Absorb: '<span class="badge text-bg-light">Absorb</span>',
+  Drain: '<span class="badge text-bg-warning">Drain</span>',
+  Selective: '<span class="badge text-bg-warning">Selective</span>',
+  Unabsorb: '<span class="badge text-bg-light">Unabsorb</span>',
+  Heal: '<span class="badge text-bg-success">Heal</span>',
+  Shrink: '<span class="badge text-bg-info">Shrink</span>',
+  Grow: '<span class="badge text-bg-info">Grow</span>',
+  'Size Steal': '<span class="badge text-bg-info">Size Steal</span>',
+  'Encase In Egg': '<span class="badge text-bg-primary">Encase In Egg</span>',
+};
+
+export const ItemModeSpan = {
+  Hold: '<span class="badge text-bg-secondary">Item: Hold</span>',
+  'Digest (Food Only)':
+    '<span class="badge text-bg-danger">Item: Digest (Food Only)</span>',
+  Digest: '<span class="badge text-bg-danger">Item: Digest</span>',
+};
+
+export const AddonIcon = {
+  Numbing: '',
+  Stripping: '',
+  'Leave Remains': '',
+  Muffles: 'bi-volume-mute',
+  'Affect Worn Items': '',
+  'Jams Sensors': 'bi-wifi-off',
+  'Complete Absorb': '',
+};
