+++ tgui/packages/tgui/layouts/NtosWindow.jsx
@@ -6,7 +6,7 @@
 
 import { resolveAsset } from '../assets';
 import { useBackend } from '../backend';
-import { Box, Button } from '../components';
+import { Box, Button, Image } from '../components';
 import { Window } from './Window';
 
 export const NtosWindow = (props) => {
@@ -47,7 +47,7 @@ export const NtosWindow = (props) => {
           <div className="NtosHeader__right">
             {PC_programheaders.map((header) => (
               <Box key={header.icon} inline mr={1}>
-                <img
+                <Image
                   className="NtosHeader__icon"
                   src={resolveAsset(header.icon)}
                 />
@@ -55,7 +55,7 @@ export const NtosWindow = (props) => {
             ))}
             <Box inline>
               {PC_ntneticon && (
-                <img
+                <Image
                   className="NtosHeader__icon"
                   src={resolveAsset(PC_ntneticon)}
                 />
@@ -63,7 +63,7 @@ export const NtosWindow = (props) => {
             </Box>
             {!!(PC_showbatteryicon && PC_batteryicon) && (
               <Box inline mr={1}>
-                <img
+                <Image
                   className="NtosHeader__icon"
                   src={resolveAsset(PC_batteryicon)}
                 />
